requirement.txt
celery==4.1.0
Django==1.10
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
MySQL-python==1.2.5
mysqlclient==1.3.7
six==1.11.0

